# ✨ feature: Complete Java Virtual Threads performance analysis with comprehensive benchmarks