--- a/src/main/java/edu/renata/fraga/java_virtual_threads_sample/JavaVirtualThreadsSampleApplication.java
+++ b/src/main/java/edu/renata/fraga/java_virtual_threads_sample/JavaVirtualThreadsSampleApplication.java
@@ -1,0 +1,13 @@
+package edu.renata.fraga.java_virtual_threads_sample;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JavaVirtualThreadsSampleApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(JavaVirtualThreadsSampleApplication.class, args);
+    }
+
+}