--- a/src/main/java/edu/renata/fraga/java_virtual_threads_sample/config/MvcVirtualThreadsConfig.java
+++ b/src/main/java/edu/renata/fraga/java_virtual_threads_sample/config/MvcVirtualThreadsConfig.java
@@ -1,0 +1,37 @@
+package edu.renata.fraga.java_virtual_threads_sample.config;
+
+import org.springframework.boot.web.embedded.tomcat.TomcatProtocolHandlerCustomizer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.core.task.AsyncTaskExecutor;
+import org.springframework.core.task.support.TaskExecutorAdapter;
+import org.springframework.lang.NonNull;
+import org.springframework.web.servlet.config.annotation.AsyncSupportConfigurer;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.concurrent.Executors;
+
+/**
+ * Configuração para MVC com Virtual Threads
+ */
+@Configuration
+@Profile("mvc-virtual")
+public class MvcVirtualThreadsConfig implements WebMvcConfigurer {
+
+    @Bean
+    public TomcatProtocolHandlerCustomizer<?> protocolHandlerVirtualThreadExecutorCustomizer() {
+        return protocolHandler -> protocolHandler.setExecutor(Executors.newVirtualThreadPerTaskExecutor());
+    }
+
+    @Bean("virtualThreadTaskExecutor")
+    public AsyncTaskExecutor applicationTaskExecutor() {
+        return new TaskExecutorAdapter(Executors.newVirtualThreadPerTaskExecutor());
+    }
+
+    @Override
+    public void configureAsyncSupport(@NonNull AsyncSupportConfigurer configurer) {
+        configurer.setTaskExecutor(applicationTaskExecutor());
+        configurer.setDefaultTimeout(30_000);
+    }
+}