--- a/src/main/resources/application-mvc-virtual.properties
+++ b/src/main/resources/application-mvc-virtual.properties
@@ -1,0 +1,11 @@
+# Configuração para Spring MVC com Virtual Threads
+spring.application.name=java-virtual-threads-sample-mvc-virtual
+logging.level.root=INFO
+logging.pattern.console=%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n
+
+# Configuração específica para Virtual Threads
+server.tomcat.threads.max=200
+server.tomcat.threads.min-spare=10
+
+# Message personalizada
+demo.config.description=Spring MVC com Virtual Threads habilitadas