… .github/copilot-instructions.md
… .github/prompts/gh_br_pr.prompt.md
… BENCHMARK_RESULT.md
… README.md
… gradle/wrapper/gradle-wrapper.jar
… gradlew
… performance-benchmark.sh
… performance-report-20250815-175328.txt
… src/main/java/edu/renata/fraga/java_virtual_threads_sample/controller/MainController.java
… src/main/resources/application-mvc-traditional.properties
… src/main/resources/application-webflux-traditional.properties
… src/main/resources/application-webflux-virtual.properties
… src/main/resources/application.properties
… src/test/java/edu/renata/fraga/java_virtual_threads_sample/JavaVirtualThreadsSampleApplicationTests.java