## 🔧 What was done?
Complete Java Virtual Threads performance analysis with comprehensive benchmarks and technical documentation.

### Changes implemented:
- Spring Boot application with 4 test scenarios comparing Virtual Threads vs traditional approaches
- Automated benchmark infrastructure with Makefile and performance testing scripts
- Comprehensive technical analysis demonstrating why WebFlux + Virtual Threads can be counterproductive
- Complete documentation with usage guidelines and performance interpretation

## 🎯 Why was it done?
Performance analysis to understand the impact of Virtual Threads in different Spring architectures.

## 🧪 How to test?
1. Quick validation: ./performance-benchmark.sh quick
2. Full benchmark: make benchmark
3. View results: make report

## 📸 Evidence of functionality
- Spring MVC + Virtual Threads: 87.5% improvement (10,308ms → 5,494ms)
- Spring WebFlux + Virtual Threads: 48.4% degradation (10,206ms → 15,144ms)

Key Insight: Virtual Threads excel for blocking I/O but can interfere with already-optimized reactive architectures.